restructure code to use the same base, extract functions to kaplay core directory
