+++ src/core/kaplay/gameObjects/index.js
@@ -1,0 +1,17 @@
+import { k } from '../../../kplayCtx';
+
+
+export const addGameObjects = (gameObjects, map, spawnpoints) => {
+  return gameObjects.reduce((gameObjAcc, cb) => {
+    const temp = cb(k, map, spawnpoints);
+
+    if (Array.isArray(temp)) {
+      gameObjAcc.push(...temp);
+      return gameObjAcc;
+    }
+
+    gameObjAcc.push(temp);
+
+    return gameObjAcc;
+  }, []);
+};
