+++ src/gameObjects/map_arcade/index.js
@@ -1,5 +1,3 @@
-import { k } from '../../kplayCtx';
-
 import { jessie } from './jessie.gameObject';
 
 import { npcsInArcadeMap } from './npcsOnmap';
@@ -10,17 +8,4 @@ const gameObjects = [
     jessie,
 ];
 
-export const addGameObjects = (map, spawnpoints) => {
-    return gameObjects.reduce((gameObjAcc, cb) => {
-        const temp = cb(k, map, spawnpoints);
-
-        if (Array.isArray(temp)) {
-            gameObjAcc.push(...temp);
-            return gameObjAcc;
-        }
-
-        gameObjAcc.push(temp);
-
-        return gameObjAcc;
-    }, []);
-};
+export default gameObjects;
