+++ src/gameObjects/map_campus_house_1/index.js
@@ -1,20 +1,6 @@
-import { k } from '../../kplayCtx';
 
 const gameObjects = [
     // Add more game objects here
 ];
 
-export const addGameObjects = (map, spawnpoints) => {
-    return gameObjects.reduce((gameObjAcc, cb) => {
-        const temp = cb(k, map, spawnpoints);
-
-        if (Array.isArray(temp)) {
-            gameObjAcc.push(...temp);
-            return gameObjAcc;
-        }
-
-        gameObjAcc.push(temp);
-
-        return gameObjAcc;
-    }, []);
-};
+export default gameObjects;
