+++ src/gameObjects/map_city/index.js
@@ -1,4 +1,3 @@
-import { k } from '../../kplayCtx';
 
 import { npcsInCityMap } from './npcsOnmap_city';
 const gameObjects = [
@@ -7,17 +6,4 @@ const gameObjects = [
 
 ];
 
-export const addGameObjects = (map, spawnpoints) => {
-    return gameObjects.reduce((gameObjAcc, cb) => {
-        const temp = cb(k, map, spawnpoints);
-
-        if (Array.isArray(temp)) {
-            gameObjAcc.push(...temp);
-            return gameObjAcc;
-        }
-
-        gameObjAcc.push(temp);
-
-        return gameObjAcc;
-    }, []);
-};
+export default gameObjects;
