+++ src/gameObjects/map_forest/index.js
@@ -1,20 +1,6 @@
-import { k } from '../../kplayCtx';
 
 import { butterfly } from './butterfly.gameObject';
 
 const gameObjects = [butterfly];
 
-export const addGameObjects = (map, spawnpoints) => {
-    return gameObjects.reduce((gameObjAcc, cb) => {
-        const temp = cb(k, map, spawnpoints);
-
-        if (Array.isArray(temp)) {
-            gameObjAcc.push(...temp);
-            return gameObjAcc;
-        }
-
-        gameObjAcc.push(temp);
-
-        return gameObjAcc;
-    }, []);
-};
+export default gameObjects;
