+++ src/gameObjects/map_forest_junction/index.js
@@ -1,18 +1,4 @@
-import { k } from '../../kplayCtx';
-
 const gameObjects = [];
 
-export const addGameObjects = (map, spawnpoints) => {
-    return gameObjects.reduce((gameObjAcc, cb) => {
-        const temp = cb(k, map, spawnpoints);
 
-        if (Array.isArray(temp)) {
-            gameObjAcc.push(...temp);
-            return gameObjAcc;
-        }
-
-        gameObjAcc.push(temp);
-
-        return gameObjAcc;
-    }, []);
-};
+export default gameObjects;
