+++ src/gameObjects/map_start/index.js
@@ -1,5 +1,3 @@
-import { k } from '../../kplayCtx';
-
 import { bruno } from './bruno.gameObject';
 import { cake } from './cake.gameObject';
 import { npcsOnChairs } from './npcsOnChairs.gameObject';
@@ -12,7 +10,7 @@ import { jokeTellerNPC } from './jokeTellerNPC.gameObject';
 import { tvVideo } from './tv_main_room_video.gameObject';
 import { randNpcsOnRestroomSinkCounch } from './randNpcsOnRestroomSinkCounch.gameObject';
 
-const gameObjects = [
+export const gameObjects = [
     bruno,
     npcsOnChairs,
     ztmTrailer,
@@ -25,17 +23,4 @@ const gameObjects = [
     tvVideo,
 ];
 
-export const addGameObjects = (map, spawnpoints) => {
-    return gameObjects.reduce((gameObjAcc, cb) => {
-        const temp = cb(k, map, spawnpoints);
-
-        if (Array.isArray(temp)) {
-            gameObjAcc.push(...temp);
-            return gameObjAcc;
-        }
-
-        gameObjAcc.push(temp);
-
-        return gameObjAcc;
-    }, []);
-};
+export default gameObjects;
