+++ src/interactions/map_arcade/index.js
@@ -1,5 +1,3 @@
-import { k } from '../../kplayCtx';
-
 import { enterMapCityInteraction } from './enterMapCity.interactions';
 import { interactionWithGameMachine8 } from './game_machine_8';
 import { interactionWithGameMachine2 } from './game_machine_2.interactions';
@@ -24,9 +22,4 @@ const interactions = [
     interactionWithGameMachine11,
 ];
 
-export const attachInteractions = (gameObjTag) => {
-    const map = k.get('main_map')[0];
-    const gameObj = k.get(gameObjTag)[0];
-
-    interactions.forEach((cb) => cb(gameObj, k, map));
-};
+export default interactions;
