+++ src/interactions/map_campus_house_1/index.js
@@ -1,5 +1,3 @@
-import { k } from '../../kplayCtx';
-
 import { enterMapCityInteraction } from './enterMapCity.interactions';
 import { bedInteractions } from './bed.interaction';
 
@@ -9,9 +7,4 @@ const interactions = [
     bedInteractions,
 ];
 
-export const attachInteractions = (gameObjTag) => {
-    const map = k.get('main_map')[0];
-    const gameObj = k.get(gameObjTag)[0];
-
-    interactions.forEach((cb) => cb(gameObj, k, map));
-};
+export default interactions;
