+++ src/interactions/map_city/index.js
@@ -1,5 +1,3 @@
-import { k } from '../../kplayCtx';
-
 import { boundaryArcadeInteraction } from './boundaryArcade.interaction';
 import { boundaryBurgerBarInteraction } from './boundaryBurgerBar.interaction';
 import { enterMapCampusHouse1Interaction } from './enterMapCampusHouse1.interaction';
@@ -30,9 +28,4 @@ const interactions = [
     enterMapCampusHouse1Interaction,
 ];
 
-export const attachInteractions = (gameObjTag) => {
-    const map = k.get('main_map')[0];
-    const gameObj = k.get(gameObjTag)[0];
-
-    interactions.forEach((cb) => cb(gameObj, k, map));
-};
+export default interactions;
