+++ src/interactions/map_forest/index.js
@@ -1,4 +1,3 @@
-import { k } from '../../kplayCtx';
 
 import { enterMapForestJunctionInteraction } from './enterMapForestJunction.interactions';
 import { interactionWithTree } from './tree.interactions';
@@ -22,9 +21,4 @@ const interactions = [
     interactionWithFirepit,
 ];
 
-export const attachInteractions = (gameObjTag) => {
-    const map = k.get('main_map')[0];
-    const gameObj = k.get(gameObjTag)[0];
-
-    interactions.forEach((cb) => cb(gameObj, k, map));
-};
+export default interactions;
