+++ src/interactions/map_forest_junction/index.js
@@ -1,5 +1,3 @@
-import { k } from '../../kplayCtx';
-
 import { enterMapCityInteraction } from './enterMapCity.interactions';
 import { enterMapForestInteraction } from './enterMapForest.interactions';
 import { interactionWithFlowers } from './flowers.interactions';
@@ -15,9 +13,4 @@ const interactions = [
     interactionWithMushroom,
 ];
 
-export const attachInteractions = (gameObjTag) => {
-    const map = k.get('main_map')[0];
-    const gameObj = k.get(gameObjTag)[0];
-
-    interactions.forEach((cb) => cb(gameObj, k, map));
-};
+export default interactions;
